SnackBar and Dialog material included

+++ src/app/recipes/pages/create/create.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute, Router } from '@angular/router';
 import { switchMap } from 'rxjs/operators';
+import { ConfirmDeleteComponent } from '../../components/confirm-delete/confirm-delete.component';
 
 import { Cuisines, Ingredient, Recipe, Step } from '../../interfaces/recipes.interface';
 import { UrlImagePipe } from '../../pipes/get-url-image.pipe';
@@ -19,16 +22,19 @@ import { RecipesService } from '../../services/recipes.service';
 })
 export class CreateComponent implements OnInit {
   ngOnInit(): void {
-    if (this.router.url.includes('edit')) {
-      this.activatedRoute.params
-        .pipe(switchMap(({ id }) => this.recipesService.getRecipeById(id)))
-        .subscribe((recipe) => (this.recipe = recipe));
+    if (this.router.url.includes('create')) {
+      return;
     }
+    this.activatedRoute.params
+      .pipe(switchMap(({ id }) => this.recipesService.getRecipeById(id)))
+      .subscribe((recipe) => (this.recipe = recipe));
   }
   constructor(
     private activatedRoute: ActivatedRoute,
     private recipesService: RecipesService,
-    private router: Router
+    private router: Router,
+    private snackBar: MatSnackBar,
+    public dialog: MatDialog
   ) {}
 
   vegetarian: boolean = false;
@@ -88,17 +94,34 @@ export class CreateComponent implements OnInit {
     if (this.recipe.id) {
       this.recipesService.updateRecipe(this.recipe).subscribe((recipe) => {
         this.recipe = recipe;
+        this.showSnackBar('Recipe updated!');
       });
     } else {
       this.recipesService.addRecipe(this.recipe).subscribe((recipe) => {
         this.recipe = recipe;
+        this.showSnackBar('Recipe added!');
       });
     }
   }
 
+  showSnackBar(message: string) {
+    this.snackBar.open(message, 'Ok!', {
+      duration: 2000,
+    });
+  }
+
   deleteRecipe() {
-    this.recipesService.deleteRecipe(this.recipe).subscribe((recipe) => {
-      this.router.navigate(['/recipes/list']);
+    const dialog = this.dialog.open(ConfirmDeleteComponent, {
+      width: '250px',
+      data: { ...this.recipe }, //pasamos la destructuración del objeto ya que en otro caso se pasaría por referencia y no sería igual de seguro
+    });
+
+    dialog.afterClosed().subscribe((result) => {
+      if (result) {
+        this.recipesService.deleteRecipe(this.recipe).subscribe((recipe) => {
+          this.router.navigate(['/recipes/list']);
+        });
+      }
     });
   }
